Corrected version file not found after install by moving version.py under py4DSTEM and updating references, also updated setup.py with required packages
